requirements.txt
Django==1.7
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Naked==0.1.31
PyYAML==3.11
Scrapy==0.24.4
Twisted==14.0.2
Werkzeug==0.10.1
argparse==1.2.1
beautifulsoup4==4.3.2
cffi==0.8.6
characteristic==14.1.0
crypto==1.4.0
cryptography==0.6
cssselect==0.9.1
dateutils==0.6.6
docutils==0.12
itsdangerous==0.24
lxml==3.4.0
pyOpenSSL==0.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
pycrypto==2.6.1
python-dateutil==2.4.2
pytz==2015.2
queuelib==1.2.2
requests==2.6.2
service-identity==14.0.0
six==1.8.0
w3lib==1.10.0
wsgiref==0.1.2
zope.interface==4.1.1
